Re-organize random structure
