+++ packages/core/src/random/distributions/index.ts
@@ -1,0 +1,8 @@
+export * from './uniform';
+export * from './integer';
+export * from './normal';
+export * from './exponential';
+export * from './poisson';
+export * from './bernoulli';
+export * from './binomial';
+export * from './geometric';
